requirements.txt
pycryptodome
flask
requests
tox
gunicorn
sly
